# docs/swagger.json
{
    "schemes": [
        "http,https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "swagger test example",
        "title": "swagger test",
        "contact": {
            "name": "Skyler"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api/v1",
    "paths": {
        "/Device": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "Get Account by index",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "auth",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "index",
                        "name": "index",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"OK\"}",
                        "schema": {
                            "$ref": "#/definitions/main.account"
                        }
                    },
                    "400": {
                        "description": "{\"msg\":\"fail\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "Add Account in Slice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "auth",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "params",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.account"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"OK\"}",
                        "schema": {
                            "$ref": "#/definitions/main.account"
                        }
                    },
                    "400": {
                        "description": "{\"msg\":\"fail\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "Delete Account by index",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "auth",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "index",
                        "name": "index",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"OK\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\":\"fail\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.account": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    }
}